Stories: Move number input example to new story
